WIP: add public interface with factory methods, hide implementation
