+++ src/test/java/io/github/jopenlibs/vault/vault/api/AuthLookupWrapTest.java
@@ -2,6 +2,7 @@ package io.github.jopenlibs.vault.vault.api;
 
 import io.github.jopenlibs.vault.Vault;
 import io.github.jopenlibs.vault.VaultConfig;
+import io.github.jopenlibs.vault.VaultImpl;
 import io.github.jopenlibs.vault.json.JsonObject;
 import io.github.jopenlibs.vault.response.LogicalResponse;
 import io.github.jopenlibs.vault.vault.VaultTestUtils;
@@ -39,7 +40,7 @@ public class AuthLookupWrapTest {
     public void should_lookup_wrap_use_url_sys_wrapping_lookup() throws Exception {
         VaultConfig vaultConfig = new VaultConfig().address("http://127.0.0.1:8999")
                 .token("wrapped").build();
-        Vault vault = new Vault(vaultConfig);
+        Vault vault = new VaultImpl(vaultConfig);
         LogicalResponse response = vault.auth().lookupWrap();
 
         assertEquals(200, response.getRestResponse().getStatus());
