+++ src/test/java/io/github/jopenlibs/vault/vault/mock/MockVault.java
@@ -29,7 +29,7 @@ import static io.github.jopenlibs.vault.vault.VaultTestUtils.readRequestHeaders;
  * server.start();
  *
  * final VaultConfig vaultConfig = new VaultConfig().address("http://127.0.0.1:8999").token("mock_token").build();
- * final Vault vault = new Vault(vaultConfig);
+ * final Vault vault = new VaultImpl(vaultConfig);
  * final LogicalResponse response = vault.logical().read("secret/hello");
  *
  * assertEquals(200, response.getRestResponse().getStatus());
